+++ packages/webpack-cli/lib/webpack-cli.js
@@ -4,7 +4,6 @@ const path = require('path');
 const { program } = require('commander');
 const getPkg = require('./utils/package-exists');
 const webpack = getPkg('webpack') ? require('webpack') : undefined;
-const { merge } = require('webpack-merge');
 const { extensions, jsVariants } = require('interpret');
 const rechoir = require('rechoir');
 const { distance } = require('fastest-levenshtein');
@@ -12,7 +11,6 @@ const { options: coloretteOptions, yellow, cyan, green, bold } = require('colore
 
 const logger = require('./utils/logger');
 const capitalizeFirstLetter = require('./utils/capitalize-first-letter');
-const { cli, flags } = require('./utils/cli-flags');
 const CLIPlugin = require('./plugins/CLIPlugin');
 const promptInstallation = require('./utils/prompt-installation');
 const toKebabCase = require('./utils/to-kebab-case');
@@ -229,6 +227,204 @@ class WebpackCLI {
     }
 
     getBuiltInOptions() {
+        const minimumHelpFlags = [
+            'config',
+            'config-name',
+            'merge',
+            'env',
+            'mode',
+            'watch',
+            'watch-options-stdin',
+            'stats',
+            'devtool',
+            'entry',
+            'target',
+            'progress',
+            'json',
+            'name',
+            'output-path',
+        ];
+
+        const builtInFlags = [
+            // For configs
+            {
+                name: 'config',
+                alias: 'c',
+                type: String,
+                multiple: true,
+                description: 'Provide path to a webpack configuration file e.g. ./webpack.config.js.',
+            },
+            {
+                name: 'config-name',
+                type: String,
+                multiple: true,
+                description: 'Name of the configuration to use.',
+            },
+            {
+                name: 'merge',
+                alias: 'm',
+                type: Boolean,
+                description: "Merge two or more configurations using 'webpack-merge'.",
+            },
+            // Complex configs
+            {
+                name: 'env',
+                type: (value, previous = {}) => {
+                    // This ensures we're only splitting by the first `=`
+                    const [allKeys, val] = value.split(/=(.+)/, 2);
+                    const splitKeys = allKeys.split(/\.(?!$)/);
+
+                    let prevRef = previous;
+
+                    splitKeys.forEach((someKey, index) => {
+                        if (!prevRef[someKey]) {
+                            prevRef[someKey] = {};
+                        }
+
+                        if (typeof prevRef[someKey] === 'string') {
+                            prevRef[someKey] = {};
+                        }
+
+                        if (index === splitKeys.length - 1) {
+                            prevRef[someKey] = val || true;
+                        }
+
+                        prevRef = prevRef[someKey];
+                    });
+
+                    return previous;
+                },
+                multiple: true,
+                description: 'Environment passed to the configuration when it is a function.',
+            },
+
+            // Adding more plugins
+            {
+                name: 'hot',
+                alias: 'h',
+                type: Boolean,
+                negative: true,
+                description: 'Enables Hot Module Replacement',
+                negatedDescription: 'Disables Hot Module Replacement.',
+            },
+            {
+                name: 'analyze',
+                type: Boolean,
+                multiple: false,
+                description: 'It invokes webpack-bundle-analyzer plugin to get bundle information.',
+            },
+            {
+                name: 'progress',
+                type: [Boolean, String],
+                description: 'Print compilation progress during build.',
+            },
+            {
+                name: 'prefetch',
+                type: String,
+                description: 'Prefetch this request.',
+            },
+
+            // Output options
+            {
+                name: 'json',
+                type: [String, Boolean],
+                alias: 'j',
+                description: 'Prints result as JSON or store it in a file.',
+            },
+
+            // For webpack@4
+            {
+                name: 'entry',
+                type: String,
+                multiple: true,
+                description: 'The entry point(s) of your application e.g. ./src/main.js.',
+            },
+            {
+                name: 'output-path',
+                alias: 'o',
+                type: String,
+                description: 'Output location of the file generated by webpack e.g. ./dist/.',
+            },
+            {
+                name: 'target',
+                alias: 't',
+                type: String,
+                multiple: webpack.cli !== undefined,
+                description: 'Sets the build target e.g. node.',
+            },
+            {
+                name: 'devtool',
+                type: String,
+                negative: true,
+                alias: 'd',
+                description: 'Determine source maps to use.',
+                negatedDescription: 'Do not generate source maps.',
+            },
+            {
+                name: 'mode',
+                type: String,
+                description: 'Defines the mode to pass to webpack.',
+            },
+            {
+                name: 'name',
+                type: String,
+                description: 'Name of the configuration. Used when loading multiple configurations.',
+            },
+            {
+                name: 'stats',
+                type: [String, Boolean],
+                negative: true,
+                description: 'It instructs webpack on how to treat the stats e.g. verbose.',
+                negatedDescription: 'Disable stats output.',
+            },
+            {
+                name: 'watch',
+                type: Boolean,
+                negative: true,
+                alias: 'w',
+                description: 'Watch for files changes.',
+                negatedDescription: 'Do not watch for file changes.',
+            },
+            {
+                name: 'watch-options-stdin',
+                type: Boolean,
+                negative: true,
+                description: 'Stop watching when stdin stream has ended.',
+                negatedDescription: 'Do not stop watching when stdin stream has ended.',
+            },
+        ];
+
+        // Extract all the flags being exported from core.
+        // A list of cli flags generated by core can be found here https://github.com/webpack/webpack/blob/master/test/__snapshots__/Cli.test.js.snap
+        const coreFlags = webpack.cli
+            ? Object.entries(webpack.cli.getArguments()).map(([flag, meta]) => {
+                  if (meta.simpleType === 'string') {
+                      meta.type = String;
+                  } else if (meta.simpleType === 'number') {
+                      meta.type = Number;
+                  } else {
+                      meta.type = Boolean;
+                      meta.negative = !flag.endsWith('-reset');
+                  }
+
+                  const inBuiltIn = builtInFlags.find((builtInFlag) => builtInFlag.name === flag);
+
+                  if (inBuiltIn) {
+                      return { ...meta, name: flag, group: 'core', ...inBuiltIn };
+                  }
+
+                  return { ...meta, name: flag, group: 'core' };
+              })
+            : [];
+        const flags = []
+            .concat(builtInFlags.filter((builtInFlag) => !coreFlags.find((coreFlag) => builtInFlag.name === coreFlag.name)))
+            .concat(coreFlags)
+            .map((option) => {
+                option.help = minimumHelpFlags.includes(option.name) ? 'minimum' : 'verbose';
+
+                return option;
+            });
+
         return flags;
     }
 
@@ -578,6 +774,8 @@ class WebpackCLI {
             const isCommandHelp = options.length === 1 && !isOption(options[0]);
 
             if (isGlobalHelp || isCommandHelp) {
+                const cliAPI = this;
+
                 program.configureHelp({
                     sortSubcommands: true,
                     // Support multiple aliases
@@ -610,13 +808,15 @@ class WebpackCLI {
                         }`;
                     },
                     visibleOptions: function visibleOptions(command) {
+                        const options = cliAPI.getBuiltInOptions();
+
                         return command.options.filter((option) => {
                             if (option.hidden) {
                                 return false;
                             }
 
                             if (!isVerbose) {
-                                const foundOption = flags.find((flag) => {
+                                const foundOption = options.find((flag) => {
                                     if (option.negate && flag.negative) {
                                         return `no-${flag.name}` === option.name();
                                     }
@@ -1108,6 +1308,8 @@ class WebpackCLI {
         }
 
         if (options.merge) {
+            const { merge } = require('webpack-merge');
+
             // we can only merge when there are multiple configurations
             // either by passing multiple configs by flags or passing a
             // single config exporting an array
@@ -1158,9 +1360,9 @@ class WebpackCLI {
             return config;
         }
 
-        if (cli) {
+        if (webpack.cli) {
             const processArguments = (configOptions) => {
-                const coreFlagMap = flags
+                const coreFlagMap = this.getBuiltInOptions()
                     .filter((flag) => flag.group === 'core')
                     .reduce((accumulator, flag) => {
                         accumulator[flag.name] = flag;
@@ -1176,7 +1378,7 @@ class WebpackCLI {
 
                     return accumulator;
                 }, {});
-                const problems = cli.processArguments(coreFlagMap, configOptions, CLIoptions);
+                const problems = webpack.cli.processArguments(coreFlagMap, configOptions, CLIoptions);
 
                 if (problems) {
                     const capitalizeFirstLetter = (string) => {
